Closer to Source Book
I cleaned up and repaired this dump based on the original "Atari BASIC Source Book" by Bill Wilkinson, Kathleen O'Brien, and Paul Laughton, including many of the original typos.

I tried to make this the best canonical source for the original Atari BASIC source code. All builds pass checks.

This language can be improved by a mile. If you want to extend this, add new version conditionals, modify the link size and headers as-needed.

I'm using this as guidance to make an accurate BASIC tokenizer and executor in Python, so I'll continue to clean it up as I go.

Have fun, and Happy 50th birthday to the 6502!